added folder checkbox to templates to give a home for roll table entries that would otherwise be in un-usable templates and be an eyesore in other parts of the program.
Fixed an issue where the shop did not properly update a comboBox with recently changed templates

fixed a bug due to duplicate update_name function in the shop.

fixed an issue where vault items did not get updated names when those names changed in the shop while editing something

added quick heal button in the style of the quick damage buttons

fixed an issue with creating a group of creatures did not allow you to group only what was selected on screen.

seperated the kill and remove buttons

updated various descriptions and tooltips based on feedback

known breaking changes: the vault table now has an additional is_folder boolean field, you may have to add this optional field to any database you wish to be compatible going forward. Or simply export the database, delete it, start the app, close the app, and import the database